Splitting the code into separate modules. Moving Currencies from attributes to global namespaces
