+++ pycbrf/exceptions.py
@@ -1,4 +1,3 @@
-
 class PycbrfException(Exception):
     """Base exception."""
