+++ pycbrf/toolbox.py
@@ -1,4 +1,6 @@
 # Exposed as API
 from .banks import Banks  # noqa
+from .currencies import Currency, Currencies  # noqa
 from .exceptions import PycbrfException, CurrencyNotFound, WrongArguments, ExchangeRateNotFound  # noqa
-from .rates import Currency, Currencies, ExchangeRates, ExchangeRateDynamics  # noqa
+from .rate_dynamics import ExchangeRateDynamics  # noqa
+from .rates import ExchangeRates  # noqa
